djustments to mapgen
-added not_badland_soils_on and badland_soils_on groups
-Reduced wetland: tanai to .65 from .8 (fix no staircase land)
-Reduced rich forest: damo to .65 from .8 (fix no staircase land)
-Added merki spawning in rich forest also
-Added Duneland: alaf spawn rate above y=100
-Lowered Barrenland: Tashvish - too much grass for barren lands
-Increased spawn rate of Jogalan
-Lowered moss spawn rate in badlands
-Added spawning of kubwakubwa in badlands
-Reduced  spawning of pegasuns in badlands - too many before
-Increased spawning of bugs in badlands.

The overall goal was to make barren lands look more realistic while
making it possible to survive if you decided to make it your home.
Found there were too many chickens and too much grass growing for an
area starved of water.  The tsaplops are intended to be added along
with this adjustment to make barrenlands more survivable.  You'll have
to eat bugs and spiders and farm tsaplops, jogalan, and tashvish,
but I think its now possible to survive using nothing but what spawns
in barren lands.

Note: there's a change to the kubwakubwa forest and Orom, but think its
just
spacing changes so safe to ignore.

I was thinking we should change the kubwakubwa_eggs_forest to _land or
something instead. Not sure how that magic works yet so didn't want to
change anything that could
break it.

I also removed the code preventing kubwakubwa from spawning in light.
Didn't see any beach spiders so think that was fixed with mapgen
updates.